requirements.txt
Flask==1.0.2
Flask-Cors==3.0.6
Flask-PyMongo==0.5.1
Flask-restplus==0.11.0
gunicorn==19.8.1
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
lazy-object-proxy==1.3.1
py==1.5.3
pylint==1.8.4
pymongo==3.6.1
pytest==3.5.1
python-dotenv==0.8.2
requests==2.18.4
urllib3==1.22
Werkzeug==0.14.1
Flask-Cors==3.0.6
PyJWT==1.4.0
coverage==4.5.1
coveralls==1.3.0
